Command Lexer (#1)
The command lexer iterates over a command file (or command string) and produces an Iterator of Commands. Each command contains one or more tokens that make up the command. The Lexer has the following capabilities:

- Expand variables (e.g., $foo)
- Expand position arguments (e.g., $1)
- Expand escaped characters (e.g., '\n')
- Handle multi-line commands
- Handle multiple word tokens using quotes (e.g., "foo bar" is one token with the quotes, two tokens without the quotes)
- Removes all extra whitespace

Also included is a basic Command Parser that was just used to play with the interface. This is expected to change as we work on that component next.