Command values (#2)
Conversions to/from value types as represented natively in Rust (i.e., i64) and in the command system (i.e., CommandValue).  For example, there are trait implementations for many types:

- let a: String = "foo".to_command_value()
- let b: i32 = 42.to_command_value()
- a.from_command_value().unwrap() -> "foo"
- b.from_command_value().unwrap() -> 42

Command values are useful because we'll need a way to pass in parameters of an unknown native type into functions which are not known by the caller.