Oso/polar integration
Integrated oso/polar to support "runtime" access of structs.
This will be useful going forward for the command system which needs
to be able to access attributes and execute methods on instances
based on a text string.

I've included the oso source here because we need access to the
runtime system, but not their text interpreter and they do not
expose the runtime system separately.

More information on the project is available here:
https://github.com/osohq/oso