wip (#4) Command Registry
The Oso files were pulled out of a separate project to make refactoring easier in the IDE and now reside in core::command::oso. The oso-derive project is still separate as required for macros. Eventually, this project will be better integrated with the rest of the codebase.

A new Command Registry struct will be the only way that users will interact with the command system.
- A facade that uses Polar under the hood for "runtime" execution
- Creates instances of structs from a String struct name lookup and PolarValues
- Converts Strings to PolarValues for the instantiation of structs and the execution of methods (candidate to move external when we need more flexibility)
- Manages the command file directory structure where instances of structs are stored

Additionally, mod.rs files are used instead for simpler organization.