Retrieve Objects and Attributes (#5)
Objects can now be retrieved from the registry in a type-safe manner.

   let user = Registry::get_instance::<User>("/foo/bar").unwrap()

Classes added to the registry will also add the classes attributes (and methods which will be enabled in a future commit) as child nodes so they can be invoked through the registry through a path. For example.

    #[derive(Clone, PolarClass, Default)]
    struct User2 {
        pub username: String,
        #[polar(attribute)]
        pub user_id: i32
    }

If User2 is registered at /foo then user_id can be obtained from /foo/user_id. The values can be accessed with a type-safe method:

        registry.make_instance("/foo",  ".", "User", &vec!["jim", "42"]).unwrap();
        let result = registry.get_attr("/foo/user_id").unwrap(); 
        assert_eq!(42, result);